feat(mock): add mock server
It's completely based on webpack-dev-server, so when you start the front-end service, the
mock-server will start automatically. Here we also observe the change of the content of the mock
folder through chokidar. When changes occur, the previously registered mock-api interface is cleared
and the new interface is dynamically mounted to support hot updates.